improve stability, UX

+++ kamatera/dto.go
@@ -6,23 +6,23 @@ type listServersPostValues struct {
 }
 
 type createServerPostValues struct {
-	Name             string  `json:"name"`
-	Password         string  `json:"password"`
-	PasswordValidate string  `json:"passwordValidate"`
-	SSHKey           string  `json:"ssh-key"`
-	Datacenter       string  `json:"datacenter"`
-	Image            string  `json:"image"`
-	CPU              string  `json:"cpu"`
-	RAM              float64 `json:"ram"`
-	Disk             string  `json:"disk"`
-	DailyBackup      string  `json:"dailybackup"`
-	Managed          string  `json:"managed"`
-	Network          string  `json:"network"`
-	Quantity         string  `json:"quantity"`
-	BillingCycle     string  `json:"billingcycle"`
-	MonthlyPackage   string  `json:"monthlypackage"`
-	PowerOn          string  `json:"poweronaftercreate"`
-	ScriptFile       string  `json:"script-file"`
+	Name             string `json:"name"`
+	Password         string `json:"password"`
+	PasswordValidate string `json:"passwordValidate"`
+	SSHKey           string `json:"ssh-key"`
+	Datacenter       string `json:"datacenter"`
+	Image            string `json:"image"`
+	CPU              string `json:"cpu"`
+	RAM              int    `json:"ram"`
+	Disk             string `json:"disk"`
+	DailyBackup      string `json:"dailybackup"`
+	Managed          string `json:"managed"`
+	Network          string `json:"network"`
+	Quantity         string `json:"quantity"`
+	BillingCycle     string `json:"billingcycle"`
+	MonthlyPackage   string `json:"monthlypackage"`
+	PowerOn          string `json:"poweronaftercreate"`
+	ScriptFile       string `json:"script-file"`
 }
 
 type powerOperationServerPostValues struct {
@@ -31,13 +31,13 @@ type powerOperationServerPostValues struct {
 }
 
 type configureServerPostValues struct {
-	ID             string  `json:"id"`
-	CPU            string  `json:"cpu"`
-	RAM            float64 `json:"ram"`
-	DailyBackup    string  `json:"dailybackup"`
-	Managed        string  `json:"managed"`
-	BillingCycle   string  `json:"billingcycle"`
-	MonthlyPackage string  `json:"monthlypackage"`
+	ID             string `json:"id"`
+	CPU            string `json:"cpu"`
+	RAM            int    `json:"ram"`
+	DailyBackup    string `json:"dailybackup"`
+	Managed        string `json:"managed"`
+	BillingCycle   string `json:"billingcycle"`
+	MonthlyPackage string `json:"monthlypackage"`
 }
 
 type changePasswordServerPostValues struct {
